7X faster bcm_tests
Summary:
Mostly bring back pieces of code retired in D7996069,D7888008.
The idea is to not do full unit initialization for each bcm_test.
This saves time for each test setting up the unit, attaching to it etc.
We still do cold boot on each bcm_test, but the time savings on
not reattaching to the unit are substantial. Concretely
99 tests take 540secs with the earlier approach and 76secs with
the new P59795462. So we could scale to ~700 tests with the similar
time incurred as done by bcm_tets today.
TODO: We need to explicitly free flex counters if we are not recreating
unit on each test. Figure out how to do this.

Reviewed By: capveg

Differential Revision: D8763960

fbshipit-source-id: 7f002b934a91ebb40dbf47914d8459bdb9c1425d