Only create LacpControllers for ports configured for aggregation
Summary:
In designing LACP initially, I had decided to create LacpControllers
not only for each physical port configured as a member of an AggregatPort, but
also for physical ports not configured as members of any AggregatePort.

The observable behavior of an LacpController for a port not configured as a
member of any bundle is as follows: 1. May only transmit an LACP frame in
response to having received an LACP frame 2. Always has its AGGREGATABLE bit off

Various network engineers have found this behavior surprising. To address this,
this diff undoes this design decision by only creating LacpControllers for ports
configured for aggregation.

Differential Revision: D8853780

fbshipit-source-id: aa212119b62cc03017e3328090d8ca0e6ba81c68