wedge_agent config: static route deletes do not take effect
Summary:
While working on D6649378, we noticed a bug in the implementation of
applyConfig().  If a static route is deleted in the config, upon a warm
boot, this deletion is not taking effect.  In other words, when wedge_agent
comes up, the deleted static route is still in the route table.

Upon investigation, the bug is the way we apply static config.  We rely on
"prevCfg" to tell us what we had applied before in order to delete stale
routes.  However, in the case of warm boot there is no prevCfg, and we
cannot reconcile stale config.

The solution is to effectively "sync" all static routes, the way we do it
for BGP. I.e., delete all existing routes and add all new routes.

Differential Revision: D8219728

fbshipit-source-id: bc92e8c263541e55cef7b16111a7e3402da86eda