Ignore Neighbor Solicitations ingressing on unbundled members
Summary:
If a (physical) port is configured as a member of an AggregatePort
(port-channel) and is physically up, but is not programmed as a member of that
AggregatePort at run-time, ARP/NDP resolution for the aggregate interface can
occur over the unbundled port.

As a direct consequence, it is possible for the IP address of the port-channel
interface to be resolved to the physical port.

This can manifest itself in various ways, the most odious of which is that when
a port-channel does not form, a single port can become fully functional in its
place.

To remedy this, ARP/NDP handling is modified to drop ARP requests and NDP
neighbor solicitations having ingressed on a port not programmed to forward
traffic as a member of its AggregatePort (if applicable).

This behavior corresponds to the intuition that a member port is functionally up
if and only if LACP has determined it should be a member of its aggregate. In
the long term, it would behoove us to stand up abstractions that make this
principle more tangibly expressed and enforced by design.

Reviewed By: boryas

Differential Revision: D8831877

fbshipit-source-id: 513f459f70f8d8b610ee1f9516a5e68cc5b54f7c