Linearize AggregatePort neighbor entry purging
Summary:
The following conditional guards the purging of ARP/NDP entries for
neighbors connected via an AggregatePort:

aggPort->forwardingSubportCount() < aggPort->getMinimumLinkCount()  (*)

MinimumLinkCount is a configuration-time parameter and so is effectively a
run-time constant. In contrast, forwardingSubportCount() is a dynamic property
of an AggregatePort. It is the number of member ports determined by LACP to
forward traffic as part of an AggregatePort.

Assume P is a member port of an AggregatePort A and consider the following sequence of events:

1. P's link state transitions to DOWN
2. NeighborUpdater is notified of state change in (1) and schedules a lambda on
the background thread that will purge A's neighbor entry if (*) is satisfied
3. LinkAggregationManager is notified of state change in (1) and (transitively)
enqueues a state update to set P's forwarding state to ForwardingState::DISABLED

If the lambda in (2) is executed prior to the state update in (3) having been
published, then the conditional in (*) will have "missed out" on the link event
in (1).

To remedy this, the decision to purge an ARP/NDP entry for a neighbor reachable
over an AggregatePort occurs when the AggregatePort transitions from UP to DOWN
in the SwitchState.

Differential Revision: D8921050

fbshipit-source-id: 17cd2148ce1531bdcfe49104e63c44f76ccee1e0