Set Interface Mode before calling SetSpeed
Summary:
When the port state is down, even if the curSpeed and desiredSpeed are same
we are calling setInterfaceMode() and setSpeed().

However, In setInterfaceMode(), we do update the mode only when the desiredMode
is different from curMode. In setSpeed() is called without setting interface
mode via setInterfaceMode(), interface mode is being set to SR4. In this change
setInterfaceMode if port is down or desiredMode is different from curMode.

Reviewed By: aeckert

Differential Revision: D9113733

fbshipit-source-id: 3020142e6444cfb3802a35d6acce8ea7615ccc25