Fix syncStaticRoutes to include default interface routes as well
Summary:
In D8219728, syncStaticRoutes() was introduced to keep the static routes updated across warmboot.

However, In syncStaticRoutes() we are getting the routeTable entries from orig_ which would not have the updated interface route entries. Ensure these entries are also part of the newly synced routes by passing the interface routes as well.

Reviewed By: tfangit

Differential Revision: D9125697

fbshipit-source-id: dc2d40f6362b629ea4c4a4ac0eb9a074bb2cd28d