Create BcmTrunk w/o specifying id
Summary:
Within new SDK6.5.13, we started seeing invalid parameter error when we tried to create bcm trunk. We used port-channel key as the trunk id(i.e, 200) to create a bcm trunk. But in the new SDK 6.5.13, for TH3,
the maximum number of trunks is 144, which is exactly the same size of logical ports in TH3.

Therefore, to avoid stricting our user to use a small value for port-channel key, we can generate the trunk id in our side before calling BRCM api.

Reviewed By: SamanKazemkhani

Differential Revision: D8956886

fbshipit-source-id: d3ca8accb0b3bb7b48696ae1f3b91095e86fad9e