Verify topo state before and after testrun
Summary:
With current approach - idea is to maintain status of topology health as class variable.
In reality that is not happening and for every instance of FbossBaseSystemTest TopoState is getting marked as healthy, resulting in 1 test failure causing rest of tests to fail instead of skip.

I tried various approaches of creating class methods , properties to handle mutation but none of those are helping.
While I continue to investigate more creating a verify_topology_state method during setUp and tearDown.

Reviewed By: jhshi

Differential Revision: D9357832

fbshipit-source-id: f4647bb78fe6f9c26d56283c43c86fb8f0bb3cdf