fix NDPOptionsHdr to check length, not data
Summary:
We recently added more generic code for parsing ndp
options. When deciding if we should keep parsing or not, we check the
data() function of the Cursor object to see if there is anything left
to read. However, data() is just a pointer to the current index in to
the buffer and always will be non-null. The result is our logs are
inundated with "NDP Option is not present" messages.

This fixes the check to instead look at length(), which gets rid of
the log messages.

Reviewed By: ashikratnani

Differential Revision: D9510765

fbshipit-source-id: cab8b952be06c409bb7dbd580ef1e26c67f30143