potential fix for segfault in getRouteTable
Summary:
I don't have a repro for this problem yet, but just
eyeballing the code this looks wrong. We take a reference to an
internal switch state node and then iterate over it. Meanwhile, we
don't have anything holding the root of the applied switch state,
meaning it could get garbage collected during this routine.

This pulls the applied state in to a shared pointer before iterating
so we know the pointer refcount is >= 1 throughout the rest of the
function.

Differential Revision: D9616220

fbshipit-source-id: dca144ecb31c41d26a51330ef0a55c167b5cc32c