Break up linkscan callback into a top-half and a bottom-half
Summary:
The upgrade to SDK 6.5.13 brought with it persistent deadlocks between
the update thread and the linkscan thread in the event of links going down.

Although all cores from these deadlocks have the same signature, the exact
locking sequence leading to the deadlock has not been identified due to GDB's
backtrace command failing with "previous frame inner to this frame (corrupt
stack?) " in one of the deadlocked threads.

However, by leveraging log lines, the deadlock can be seen to occur while the
update thread is executing bcm_trunk_member_delete() and the linkscan thread
bcm_trunk_bitmap_expand().

Broadcom has reported that an application calling back into the SDK in the
linkscan callback context is unsupported.

With this in mind, the linkscan callback is being split into a top-half and a
bottom-half.

In analogy to the Linux kernel's facility for servicing interrupts, the top-half
will delegate the brunt of the work required to service the link-event to the
bottom-half. The top-half will be executed inline in the linkscan callback
whereas the bottom-half will be executed asynchronously on a dedicated thread.

This avoids a perversion in locking order because locks are released in the
top-half and are only reacquired via the bottom-half.

This functionality is currently guarded by a enable_linkscan_bottom_half
gflag. Special attention has been paid to avoid altering the linkscan callback
path when the gflag is not supplied.

Reviewed By: tfangit

Differential Revision: D9388674

fbshipit-source-id: 64722d77cd7218412af08586b87f9f1059e24464