Down all ports should toggle and check admin state not oper state
Summary:
DownAllPorts context mgr is supposed to admin down all
ports and then admin enable them on exiting. However, it was using
oper state as a guide to
a) Decide which ports to down - So if a port was oper state down,
we would skip disabling it. Bad because a enabled port could come
up in the middle of a test and cause noise
b) Looking at oper state post test to see if we can get a previously
downed port back up. This is again subject to all kinds of extraneous
link failures that may occur in the span of a test

Reviewed By: capveg

Differential Revision: D9800322

fbshipit-source-id: 1615f2c613cc1a541fd28726f936ce93ce78dac1