Tag bmc cmdline tests as new-test
Summary:
As it stands we want to use trunk-stable tests
as a input to health of agent. bmc cmdline tests don't provide
meaningful input to that. So for now, marking them as new
test, as we evolve we might need to tag tests based on
the binary they provide signal for

Differential Revision: D9844490

fbshipit-source-id: 581e5232a33487247b1ed657f7c61213ab5f8f1d