Use BassetButler to make reservations
Summary:
I updated the FBOSS system test code to use the BassetButler instead of getting reservations from Flitr.

There is still some additional setup code to add the reservation to Flitr because Flitr is still being used for SSH and SCP.

I was able to deprecate some unused methods. I also moved the health check functions to a separate file to improve the readability of the main file.

Differential Revision: D9926942

fbshipit-source-id: 75e93c6d91e8716dc13a56a600dd447dbc225c8a