Relax constraint in CoppTests to avoid failing tests
Summary:
It seems like despite our best efforts to isolate Copp traffic, we're
still seeing bursts of packets from time to time that hit our Copp queues.
My current theory is that they are generated from the agent itself, but
since we've been unable to capture the packets (it happens infrequently),
I've been unable to verify this.

The diff relaxes the "make sure we don't see too many extra packets"
test fail down to a stern warning in the logs.  This should reduce
the false negative rate of the test.

Reviewed By: jhshi

Differential Revision: D9977721

fbshipit-source-id: 6df2282b9a4e9cc7921aaef6f960aaf506473861