allow multiple channels to map to single WedgePort
Summary:
Previously, our PlatformPort objects would have an optional
TransceiverID and an optional ChannelID. This presumes that every port
has either one or zero associated transceivers and one or zero
associated channels. However, that may not always be the case on new
platforms.

This introduces a struct named FrontPanelResources that is a
TransceiverID and a vector of ChannelID. This replaces the two
optional ids we had before in WedgePort and enables properly mapping
each port to the channels that make up the port.

I also added a mechanism to create the port mapping from a PortID ->
FrontPanelResources mapping, instead of just PortID ->
TransceiverID. Previous model always assumed we had the same number of
ports per transceiver on every platform, which would make it difficult
to mix different line cards.

Reviewed By: tfangit

Differential Revision: D8888678

fbshipit-source-id: f4299ba220a27ab8524ae38111fbe7426ce43c28