add SaiAttribute class for safely using sai_attribute_t
Summary:
The library attempts to make the interaction with the sai_attribute_t.value
union as nice as possible. It is documented in the code, particularly in the
docs near the SaiAttribute class templates. The high level idea is to
encapsulate the meta information about an attribute found in the SAI
documentation into aclass and then use type traits to make correct accesses
into the union.

Reviewed By: tfangit

Differential Revision: D10129981

fbshipit-source-id: 19a00fe186268b967e8a7ddf2a65b65ebe5a7439