Make call to fb303 counter retryable
Summary:
Modified all uses of getCounter() to use a retryable method. This hopefully fixed the problem with thrift calls returning 0 bytes in certain situations.

There didn't seem to be an existing file in testutils that made sense for this, so I created a new fb303_utils file. I'm open to suggestions on better places for it.

Reviewed By: jhshi

Differential Revision: D10202590

fbshipit-source-id: 794e93e6db74db917c5d7c1cbffeccb1d8ba8d46