Always do work in linkscan bottom half
Summary:
Diff stack ending in D10123137, enables testing with linkscan
bottom half set to true. There are 30+ ECMP/UCMP test that get
exercised with bcm_tests. D10123137 adds tests for trunk shrink as
well. Infact as noted in D10123137, I am seeing lockups in trunk
shrink code w/o bottom half enabled in 6.4.10 itself.
So the plan is to have this be enabled in tests for a couple of
weeks and then just have bottom half be enabled always.

Reviewed By: boryas, SamanKazemkhani

Differential Revision: D10189106

fbshipit-source-id: 2a704824e365b6cc188515b2c0ea7ca94f952016