remove configerator fetching for bcm config + move to oss
Summary:
Similar to parent diff, but we also have a remote fallback
mechanism for broadcom configs. This gets rid of that and then moves
the rudimentary class for wrapping the broadcom config to open
source. It is not actually used yet, but should be possible to support
configs in open source as well.

Differential Revision: D8382574

fbshipit-source-id: 11b2089d373d82d45e2d40d28af25f10f4632421