single agent config struct + convenience class
Summary:
This creates a new struct that we can extend to encompass all
configuration needed for the agent. Debated just adding fields to
SwitchConfig, but I think it is nice to keep chip-agnostic and
chip-specific configuration fields separated.

I also pulled logic for loading/serializing/deserializing the config
out of ApplyThriftConfig and in to a helper class for more contained
access.

Differential Revision: D8382572

fbshipit-source-id: a0a88a34c74c8dd9054a12925da6170a688fe227