Fetch config on startup + move ownership to platform
Summary:
This is first step in making it possible to pull default
command line flags from the config. We need to move config parsing out
of SwSwitch::applyConfig and do it as soon as possible on startup.

The question then becomes: who owns this more generalized config?
SwSwitch used to be the owner of the config, but now that it contains
more information, it is a bit strange to store it there. Furthermore,
if we want to take advantage of the chipConfig section of the config,
we need to have access to the config during initializing the Platform
class. For these reasons, I moved ownership of the config to the
platform and modified some of the platform apis. I ended up adding
this functionality in the outermost 'Platform' class.

Differential Revision: D8382576

fbshipit-source-id: 74bc353b08a294d16a4bceabf86be284399fc02a