dump running config to set location
Summary:
After we set the running config of the agent, this dumps the
raw config to a set location. This will be useful for ensuring that
the running config matches what we expect, and can use this along w/
coop to get metadata about the config running on the box.

Reviewed By: phshaikh

Differential Revision: D8382575

fbshipit-source-id: 2362a5e2736ef44f7741985dd05dcec1651af6e0