Check qsfp string valid before return
Summary:
We occasionally saw there're some transceiver info returns invalid string to decode, mainly in Vendor field.
And python client will be not able to decode this implicitly (https://fburl.com/g85hxbil), and it's not trivial for us to fix it for thrift python client. Therefore, the better way is for the server side(our qsfp_service) should do this valid check.

Differential Revision: D10454190

fbshipit-source-id: 4cdfac6b0e09e9415ed216ea21504746f9eb5efc