add support for sai apis with members
Summary:
Some sai apis concern objects which have sub-objects which
tend to be called members. (not always, but we will use the name anyway)
For example we use the primary vlan apis (create/remove/get_attr/set_attr)
to interact with vlans and the member counterparts to interact with vlan
members (ports in a vlan). Likewise, for the bridge api, bridge_port is
a member of a bridge.

Differential Revision: D10253717

fbshipit-source-id: d3ca59c64998cf77a76039be1f0ad26625027e0b