add sai bridge api
Summary:
sai_bridge_api_t contains functions for setting up
bridges. At the very least we need to setup a single bridge
and a bridge_port for each port we want to populate into vlans.
This is an example of a SaiApi that uses members:
bridge_ports are members of a bridge

Differential Revision: D10253765

fbshipit-source-id: 40c81985f0296afa72910a8eada65e139d5ba5ee