Decouple RouteUpdater from SwitchState
Summary:
RouteUpdater is written to operate on the RouteTableRib and Route
SwitchState nodes. RouteUpdater uses these nodes in two ways:

1) Inserting into and removing routes out of a NodeMap. This NodeMap provides
the permanent storarge for all routes in FBOSS.

2) Materializing a RadixTree (providing a mapping from an IP subnet to a Route
object) from this NodeMap in the service of recursive route resolution.

As a step in building out a stand-alone RIB, this implements a rendition of
RouteUpdater which performs both (1) and (2) on a long-lived RadixTree (of type
NetworkToRouteMap).

As before, RouteUpdater assumes ownership of the data-structures it
modifies. However, because these data-structures are no longer copy-on-write,
RouteUpdater need not implement inline clone() logic. RouteUpdater assumes its
NetworkToRouteMap member variables have been appropriately locked externally.

As a byproduct of decoupling RouteUpdater from SwitchState, multi-parameter
template methods have been simplified to only use AddressType as their sole
template parameter. This constitutes the minimal templating necessary to avoid
code duplication, and is in fact the family of types with respect to which
RouteUpdater is generic.

Unrelated to the decoupling from SwitchState, the new rendition of RouteUpdater
has been made independent of VRF.

Differential Revision: D10018613

fbshipit-source-id: 3dbb3a6ff1bb9b27c7707522ff029a1bdeae4c4c