Sync packet send API
Summary:
For dataplane bcm_tests we want to wait till packet has been
sent out before we start asserting on the resultant counters. Typically
the packet send is fast enough such that we haven't seen a problem in
tests. But it seems better to rely on a synchronous interface before
asserting for post packet send conditions.

Differential Revision: D12917337

fbshipit-source-id: f0cf2122cca4b3c7ce0d7f8d1739ca7b1bb627f4