add sai router interface api
Summary:
This api allows interacting with router interfaces. A router
interface is needed to create a neighbor or a next hop which are needed
to create routes. It can be attached to a vlan, a port, or even a sub-port.
It optionally lets you configure a source mac for the router interface.

Differential Revision: D10319310

fbshipit-source-id: f423cf5c2dfd0b336f2e4330dcd04063f0febe33