fix refreshing transceivers
Summary:
Looks like some diffs a month or two ago broke refreshing
such that we fail on the first transceiver that is missing. This makes
it so we fail to program most transceviers. Looks like there are two
regressions:

1. Somehow trying to read data from missing transceivers, which leads
to throwing an FbossError.

2. We don't catch FbossError at the outer layer in WedgeManager.

We should really clean up the exception handling here to be more
intentional and not just catching all exceptions blindly, but this
will unbreak prod.

This is likely root cause for an uptick of transceiver related tasks.

Differential Revision: D12973060

fbshipit-source-id: 0fb428311e3e98d930d2b1daa5d00e38ded3091a