wedge_qsfp_util should get DOM data from qsfp_service by default.
Summary: Since we already have a service which can provide us the data we need from the Transceiver, we updated the tool to utilize that and free us from having to transport data through the i2c bus. It will also allow us to run the tool on hosts other than the box itself if we only need to read the data. Added direct_i2c gflag to preserve the old behaviour.

Reviewed By: aeckert

Differential Revision: D12876134

fbshipit-source-id: 8d057352247d0fa32a6b18a721c8dd1def9f9106