do not assume exclusive access to bus
Summary:
We removed unselecting every module after each read/write,
but this is actually not safe to do in a world where qsfp_service can
come in and modify the state of the muxes.

This adds back in the unselect call for now and makes unselect cheaper
on the PCA9548MuxedBus implementation. Once we ensure exclusive bus
access we can potentially remove this again as an optimization.

Reviewed By: phshaikh

Differential Revision: D13002239

fbshipit-source-id: 1ac1c90f7b84d06c14b98daa55a6b5570532be32