fix enabling cdr/rate select if page 0 not selected
Summary:
This tool can be used to enable/disable cdr. However, it
pretty reliably failed for me when I tried to use it.

Reason is that it assumes that page0 is selected on the transceiver
when determining if CDR is supported, which is based on a field in
page0. Most of the time page3 is selected in our dcs because we loop
through each transceiver, scan in page 0, then scan in page 3, never
resetting back to page 0.

This makes commands that need to look at page0 to actually ensure
page0 is selected before reading in the field. This applies to CDR +
rate selection.

Reviewed By: phshaikh

Differential Revision: D12969311

fbshipit-source-id: a310f6c75364723dd80abdd3f5e3da798e715770