Back out "[fboss][agent] Name the queues"
Summary:
Some of the prod switches are upgraded to SDK 6.5.13 before an issue was discovered: S168777. It is fixed by D13036591. We would like the fix to go out immediately.

There are a couple of issues around naming of the queues, so this patch series will revert it so that aforementioned D13036591 can go out off trunk. Will resubmit after fixing all the issues.

Original commit changeset: b579ebbfaade

Differential Revision: D13047418

fbshipit-source-id: 97c1c82c494e157ddc31d2efc41d2a798d560c59