Introduce RoutingInformationBase
Summary:
RoutingInformationBase holds all the routes in an FBOSS system.

As route resolution logic is entirely contained within RouteUpdater, and
RouteUpdater assumes it has unfettered read-write access to NetworkToRouteMap's,
RoutingInformationBase primarily serves to define the model by which concurrent
operations can act on routes. Examples of such operations include finding the
(most specific) network to which an IP endpoint belongs to and recursively
resolving routes.

Because we've yet to determine this concurrency model, this equips the
RoutingInformationBase object with a write-only public interface.

[1] Note that a next-hop may belong to a different address family than its
corresponding prefix, allowing for, say, an IPv4 network with an IPv6 next-hop.

Differential Revision: D10236028

fbshipit-source-id: 152b8c8ca95509a33df6c3114fa953af0c4e4f75