Model Forwarding Information Base as a SwitchState node
Summary:
The following nodes are added to SwitchState to track the FIB:

ForwardingInformationBase- a mapping from prefixes to Routes (templatized by
address family)

ForwardingInformationBaseContainer- a container for two instances of
ForwardingInformationBase (one for IPv6 and one for IPv4)

ForwardingInformationBaseMap- a mapping from VRFs to
ForwardingInformationBaseContainers

This layout mimics that of RouteTableMap, RouteTable, and RouteTableRib.

The Route class forms the leaf-node in the RouteTable hierarchy; each Route
object contains the set of all next-hops known for a given prefix (in addition
to the best next-hops). In the ForwardingInformationBase, a leaf-node represents
only the best-path(s) for a given prefix. To avoid modifying BcmSwitch, the
Route object from the RouteTable hierarchy is re-used as the leaf-node in the
ForwardingInformationBase hierarchy. Once the transition to a stand-alone RIB
has been completed, this Route object can be stripped of extraneous
functionality.

Differential Revision: D10474054

fbshipit-source-id: 86dcb6ab96f3eec8cbd65463f81a83fe290289e5