Forward LACP frames via strict-priority queue
Summary:
We have observed port-channels flaps in the Fabric Aggregator tier
coinciding with an increase in production traffic transiting through this
tier.

These port-channel flaps were root-caused to the following sequence of
events:

1. LACP frame injected into the ASIC from the FBOSS agent (residing on the
control processor) for transmission out of a front-panel port (and over a
port-channel)

2. LACP frame competing for _local_ MMU resources with application-layer traffic
and ultimately being dropped as a result

3. Remote LACP endpoint timing out and subsequently transmitting an LACP frame
with actor state COLLECTING & DISTRIBUTING bits reset to 0

4. Local LACP endpoint learning of the partner's state in the LACP frame from
(3) and removing the corresponding member port from the associated bundle as a
consequence

To eliminate this failure mode, LACP frames are marked with a class-of-service
value that is serviced by a strict-priority MMU queue reserved for
network-control traffic.

Reviewed By: boryas

Differential Revision: D13230394

fbshipit-source-id: 11800a8e393a0a28f49632cce0aae1da6bd50319