Support specifying CoS for transmitted frames
Summary:
If a frame is injected into a Broadcom ASIC for transmission out of a
pre-determined port, then processing by the ingress pipeline for the frame is
skipped. As a consequence, the frame does not undergo ACL processing.

Since FBOSS controls queueing by leveraging ACL actions to set class-of-service
for frames, frames originating from the FBOSS agent and destined to a
pre-determined set of front-panel port are not subject to the desired queueing
model. Instead, all such frames are serviced by the default (low-priority)
queue.

This extends the sendPacketOutOfPortAsync interface to allow for specifying
class-of-service for the frame being transmitted, which in turn allows for
locally-originated frames to be serviced by the desired queue.

Differential Revision: D13188349

fbshipit-source-id: 55af8b13985bf2cd1dcee841670f1f5aae62fdf7