Apply mirrors to acls.
Summary:
Apply mirror to ACL entries
1) During ACL creation, ACL entries try to apply mirror to themselves

This is to enable a use case where ACL entry could be modified to have mirror MatchAction or new ACL entry would be added with mirror MatchAction, while mirror is already present in configuration and is also resolved.

This will enable the traffic hitting ACL entry to be mirrored as soon as configuration is in effect

2) After mirrors are resolved, each ACL apply or unapply mirrors associated with it.

This is to enable a use case where mirror configuration or resolution parameters (tunnel attributes) change. In which case, ACL entries which are already using mirrors need to be updated to use a new/updated mirror.

All changes are symmetrical to how mirroring for ports is supported.

Differential Revision: D9846329

fbshipit-source-id: d8d9efbb29b004310de8b6a10f3f47f5dbfeb258