IPv4: fix wrong headers in the tests
Summary:
This diff is part of an effort to make our network parsers more robust and rfc compliant.

Some of the packets in our tests are malformed, the length field in the IPv4 header *must* includes the header size. The result is that we are generating packets with a payload that is too small.

The tests never catch the problem because our IP and UDP parsers never check the length of the payload and are fine reading outside of it (in the FCS field for example).

Differential Revision: D13179949

fbshipit-source-id: 386466dcf70fda55fa99b66f9d6658dbf14fcc20