Rename CPU queues, check FIB_SYNCED, retry w old names
Summary:
With diff later in this stack, we name the queues (e.g. cpuQueue-mid
instead of queue2). However, there are system tests that use old counter names.

This patch fixes the system tests to use new names. However, at this point in
time, production does not have the agent fix that names the queues. Thus, the
modified system test that looks for new names will fail against the production.

As a temporary workaround, this patch adds a decorator that retries querying
with old counter names if query with new counter name fails. Once the naming of
queues makes it to production, the part of this patch that adds retry
decorator could be removed. Sadly, since we don't want system test to fail on
any patch (even intermediate patch of stack diff) means that the retry logic
must be part of this patch and not a separate patch. This makes removing retry
logic somewhat non-trivial.

Reviewed By: jhshi

Differential Revision: D13093547

fbshipit-source-id: 79f4b87381102fc5826603ac8c90ab4dd99229f8