revamp sai api lib build strategy
Summary:
The original design for building sai/api:api wasn't actually correct:
It relied on the assumption that each cpp_binary would provide a dependency
that selected an actual SAI implementation that would then provide the SAI
symbols. This worked while sai/api:api was essentially a header only library
(with only non-SAI functionality in cpp files, like AddressUtil and std::hash).
However, as we start to build sai/switch/ we see that eventually we will have
libraries that need to be built in a standalone fashion that depend on only
sai/api:api. That can't work because there will be undefined symbols from sai.
The specific example that prompted this was creating SaiApiTable for use by
SaiSwitch.

The proposed solution to keep the flexibility of not tying sai api to any
particular sai implementation while allowing for compiling it as a library
is to parameterize sai/api:api by sai implementation. This diff demonstrates
taking sai/impl:fake, but it is easy to imagine adding other implementation
libraries to sai/impl.

This is a pretty important decision, and can be hard to fix as the project
grows and the build becomes more complex, so if people have ideas from
solving this before, I would really appreciate advice.

Differential Revision: D12842995

fbshipit-source-id: dbf620d9f3edb09a1fabf1e93e687e2376fb0602