use state validation to accept/reject mirror configuration
Summary:
Removing  validations that use magic number of 4 to accept or reject mirror configuration.

Instead of that, use existing "isValidStateUpdate" mechanism. This functions uses a constant defined in bcmswitch.thrift to validate if number of mirrors do not exceed a limit.

Since this validation is now moved into Bcm or Hw layer, updated tests

Differential Revision: D13113327

fbshipit-source-id: 4ab4a8a45c0490c5d6a23e1f40db5ab0108a2003