QoS: support for qos policies in the config/state
Summary:
Add a new QosPolicy object to describe the mappings of a QoS qualifier (ex.: dscp) to a Cos Queue.
These policies can then be attached to the ports.

At the moment, it is only possible to define a global QoS policy that will be attached to all the ports,
but it is trivial to add a map to override the policy. This will be implemented in another diff.

Differential Revision: D12826365

fbshipit-source-id: ace4ea9082ec17c810ad69963a5e55e24255a320