support for prefixes in SAI AddressUtil
Summary:
Routes use sai_ip_prefix_t which is a struct consisting of an address
family and two unions representing ip addresses. In future versions of SAI, this
changes to simply a pair of sai_ip_address_t structs, which seems better. It is
not clear to me why this struct is not just 'sai_ip_address_t dest; uint8_t len;'
but we will have to do some conversion.

Some other options for which prefix type to use on "our end" are the fboss prefix
type which has some more fancy functionality compared to folly::CIDRNetwork which
is just std::pair<folly::IPAddress, uint8_t>. Alternatively, we can model the sai
prefix type more closely with something like
std::pair<folly::IPAddress, folly::IPAddress>.

Finally, I don't think the algorithms I have implemented for this stuff are quite
perfect, and end up copying things around a lot. I can change the APIs to be a
more "in-place" style to improve this, if people would like me to.

Differential Revision: D10467779

fbshipit-source-id: 55e8945d13d96a9cfb14d79c45b1b335d1cbf207