Correct switch_state to use CONFIGURED instead of FIB_SYNCED
Summary:
The original copp tests need the switch state to be FIB_SYNCED, which is too strict for copp testing. To reach FIB_SYNCED also needs to depend on BGP sending us fib. We should just need to wait for switch to be CONFIGURED.
Besides I also increased the timeout to be 20 seconds.

For now, I totally replaced the old fib_synced check to configured check to avoid other people to misuse that logic.

Reviewed By: capveg

Differential Revision: D13523035

fbshipit-source-id: dc39362ab0182ba85fe0ec8416b74e962148e6ce