fboss/agent/lldp: Unit test length check too tight
Summary:
The packet length for LLDP was made to be correctly sized, but this makes
the length check in the unit test still too simplistic for all the
environments it can run in.

The test is breaking for some people.

Fix it by increasing the range of the min and maxlen checks, according to
what could be reasonably allocated.

Reviewed By: schischi

Differential Revision: D13518937

fbshipit-source-id: 235fd8e8fa1e2c6dd0b245f55c4b148c19d98822