populate mirror settings for port on warmboot
Summary:
Reconstructing mirror settings for BcmPort during warmboot.

Facebook

Provided function in BcmPort which uses warmboot cache to extract some settings which depend on dumped switch state and are required to persist across warmboot but are not present in ASIC.

This is required for mirror settings, because mirror setting in BcmPort is used in BcmPorts' destructor to unmirror a port. Not having this setting (even though port is being mirrored), will not unmirror a port.

Differential Revision: D13337703

fbshipit-source-id: 5b6616c5daa0da6207d7639d53bc162301e6a86c