Add a FakeWedge40Platform to run in custom unittests
Summary:
The only difference b/w this and a real wedge40
platform is that it writes to /tmp rather than /var/facebook/
and /dev/shm/.
There are 2 interesting decisions here
i) Creating a separate Fake platform in addition to FakeBcmTestPlatform.
This is intentional as BcmTests bring in their own quirks to query
the platform on, so keeping them separate here seems fine
ii) Creating a FakeWedge40Platform and not just a FakeWedgePlatform.
This implies we might want to distinguish b/w the platforms in emulation
the same way we do for real wedge_agent. This may or may not be true.
For the custom_unittests (coming up in subsequent diffs) it also throws
up a interesting choice, in that, should we run each test N times, once
for each platform?. These tests are pretty cheap, so it does not seem to
be a huge deal. But we will see.
For now going with the most straightforward and future flexible approach.

Differential Revision: D13579087

fbshipit-source-id: 11034136e263a9cc63b5dc47db1f20a9caca7411