add retries when setting led mode
Summary:
When starting up the agent, we make a single i2c write to the
CPLD to make sure the LED mode is set to 12-bit, not disco. However,
in cases that the i2c bus is busy during this write attempt, we raise
a LIBUSB_BUSY_ERROR and then continue starting up the agent.

This basically lifts the retry logic we use in wedge_qsfp_util and
moves it in to the agent for this write. It seems like this is a bit
more common on FA, which are more heavily loaded with optics than RSW.

Differential Revision: D13540276

fbshipit-source-id: a4f44e96f1e35b95cef3a0790bd249ada9d9272d