use lanes in FakeSaiPort; refactor FakeSaiPort to match other Fake ctors
Summary:
Now that we understand ports a bit better, it has become apparent that
lane information will be external to SAI (that is, selecting it will require
knowledgeo of a particular forwarding element's port setup). For that reason,
make the fake follow that model and just consume the lanes, like other
MANDATORY attributes in other fakes.

Reviewed By: tfangit

Differential Revision: D13116415

fbshipit-source-id: cec5255706ab0427495b5de892ce34f57b087ce5