create SaiAttributeDataTypes which share behavior with SaiAttribute
Summary:
In building SaiSwitch, it became apparent that it would be useful
to express the SaiAttributes in a given API which are needed during the
creation of an object. For example, when creating a port, we _must_ specify
the speed and list of hw lanes. Sometimes, fields will be contingent on the
values of other fields. For example, a route with packet action FORWARD
needs a next hop id, but one with packet action TRAP does not. To the end
of supporting these use cases, we add SaiAttributeTuple and
SaiAttributeOptional, which we will use to define a new create method in
SaiApi.

As for the design of the SaiAttributeDataTypes themselves: the intent is to
have them come as close as possible to implementing the same "SaiAttribute"
concept as the SaiAttributes themselves. That means they have a value and
SaiAttr method. The tricky thing is that SaiAttr could be a list of attrs,
so it will take some migrating/api changes to fully unify the classes. For
now, I have used "saiAttrV" which returns a vector of sai_attribute_t, but
I think a saiAttr() which takes an iterator to fill into would probably be
a nicer design.

To make construction of these nested, composed structures like
SaiAttributeTuple<SaiAttributeOptional<SomeAttribute>, SomeOtherAttribute>
a bit more palatable, I have added a few layers of implicit constructors.
In general, these have been pretty pleasant to work with, and make sense
since the SaiAttributes are in some sense wrapper types.

Reviewed By: tfangit

Differential Revision: D13453846

fbshipit-source-id: 91321cfc9ac3ce61abb118dec6d9c39181c7dc7d